refactor #29 : standardize DTO package naming to lowercase
- Rename RequestDto -> request, ResponseDto -> response
- Update imports in UserController, UserService, KakaoController,
  KakaoService, NaverController, NaverService, JwtAuthenticationFilter
- Follow Java package naming conventions
