+++ backend/src/main/java/com/phonebid/app/member/controller/NaverController.java
@@ -5,7 +5,7 @@ import com.phonebid.app.common.dto.ApiResponse;
 import com.phonebid.app.common.exception.CustomException;
 import com.phonebid.app.common.errorcode.NaverErrorCode;
 import com.phonebid.app.jwt.JwtUtil;
-import com.phonebid.app.member.dto.ResponseDto.LoginResponseDto;
+import com.phonebid.app.member.dto.response.LoginResponseDto;
 import com.phonebid.app.member.service.NaverService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
