+++ backend/src/main/java/com/phonebid/app/member/dto/response/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.phonebid.app.member.dto.ResponseDto;
+package com.phonebid.app.member.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
