+++ backend/src/main/java/com/phonebid/app/member/service/NaverService.java
@@ -10,8 +10,8 @@ import com.phonebid.app.jwt.JwtUtil;
 import com.phonebid.app.member.domain.Provider;
 import com.phonebid.app.member.domain.Role;
 import com.phonebid.app.member.domain.User;
-import com.phonebid.app.member.dto.ResponseDto.NaverUserInfoDto;
-import com.phonebid.app.member.dto.ResponseDto.LoginResponseDto;
+import com.phonebid.app.member.dto.response.NaverUserInfoDto;
+import com.phonebid.app.member.dto.response.LoginResponseDto;
 import com.phonebid.app.member.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
