+++ backend/src/main/java/com/phonebid/app/member/service/UserService.java
@@ -5,12 +5,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.phonebid.app.common.exception.CustomException;
 import com.phonebid.app.common.errorcode.CommonErrorCode;
-import com.phonebid.app.member.dto.RequestDto.SignupRequestDto;
-import com.phonebid.app.member.dto.RequestDto.LoginRequestDto;
-import com.phonebid.app.member.dto.RequestDto.ProfileUpdateRequestDto;
-import com.phonebid.app.member.dto.RequestDto.PasswordChangeRequestDto;
-import com.phonebid.app.member.dto.ResponseDto.LoginResponseDto;
-import com.phonebid.app.member.dto.ResponseDto.ProfileResponseDto;
+import com.phonebid.app.member.dto.request.SignupRequestDto;
+import com.phonebid.app.member.dto.request.LoginRequestDto;
+import com.phonebid.app.member.dto.request.ProfileUpdateRequestDto;
+import com.phonebid.app.member.dto.request.PasswordChangeRequestDto;
+import com.phonebid.app.member.dto.response.LoginResponseDto;
+import com.phonebid.app.member.dto.response.ProfileResponseDto;
 import com.phonebid.app.member.domain.User;
 import com.phonebid.app.member.domain.Role;
 import com.phonebid.app.member.repository.UserRepository;
