+++ backend/src/main/java/com/phonebid/app/security/JwtAuthenticationFilter.java
@@ -2,7 +2,7 @@ package com.phonebid.app.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.phonebid.app.jwt.JwtUtil;
-import com.phonebid.app.member.dto.RequestDto.LoginRequestDto;
+import com.phonebid.app.member.dto.request.LoginRequestDto;
 import com.phonebid.app.member.domain.Role;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
